Set the keymap to 'en-us' in the libvirt machine template
The setting is a bit confusing imho, but this tells libvirt that the
guest is using the en-us keyboard layout and creates a translation
table for him. This is a bit quirky since qemu has to send scancodes
which cannot be easily derived from the information vnc provides.

With this in it is possible to use alternative keyboard layouts when
connecting to the virtual machines via libvirts vnc viewer. It has
some quirks but works well for colemak and doesn't break qwerty.

Closes #5.

Signed-off-by: Justus Winter <4winter@informatik.uni-hamburg.de>